# swagger.yaml
openapi: "3.0.3"
info:
  title: Agentica Inventory API
  version: "1.0.0"
  description: |
    API REST para gestión inteligente de inventarios con búsqueda semántica y asistente IA.
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user:
          type: object
          properties:
            id:       { type: string }
            username: { type: string }
            role:     { type: string }
...
    Error:
      type: object
      properties:
        error: { type: string }
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: Iniciar sesión y obtener JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Token JWT generado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/auth/register:
    post:
      summary: Registrar nuevo usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "201":
          description: Usuario creado
        "409":
          description: Usuario ya existe
...